wsd: Rename platform & config specific files
Renamed to follow naming convention in common/.
